fix:providing new component for support modal
