+++ app/pages/ai_chat.py
@@ -1,6 +1,7 @@
 import streamlit as st
 from apis import stock_info, gemini_with_history
 from utils import session, invalid_data
+from components import support_author
 
 
 def ai_chat():
@@ -8,7 +9,7 @@ def ai_chat():
     symbol = session.set_session_state()
     stock = stock_info.StockInfo(symbol)
     stock_general_info = stock.get_general_info()
-    support_modal()
+    support_author.support_modal()
 
     user_prompt = st.text_area(
         """Ask your questions from AI based on the chosen stock symbol 
@@ -33,32 +34,6 @@ def ai_chat():
             response_placeholder.write(story)
         else:
             invalid_data.invalid_stock_symbol()
-
-
-@st.dialog("Support Us")
-def support_modal():
-    st.write(
-        """
-        <div style="display: flex; align-items: center;">
-            <system76-Pangolin-6800U-LPDDR5-Drive/dp/B0BVGJ8Y72pan style="margin-right: 1rem;">❤️</span>
-            <span>
-                Support the developer: <a href="https://buymeacoffee.com/pooria7" target="_blank">Buy me a Book</a>
-            </span>
-        </div>
-        """,
-        unsafe_allow_html=True,
-    )
-    st.write(
-        """
-        <div style="display: flex; align-items: center;">
-            <span style="margin-right: 1rem;">❤️</span>
-            <span>
-                Follow the builder on X: <a href="https://x.com/PooriaTaghdiri" target="_blank">Pooria's X</a>
-            </span>
-        </div>
-        """,
-        unsafe_allow_html=True,
-    )
 
 
 ai_chat()
