+++ app/pages/general_ai_analysis.py
@@ -1,6 +1,7 @@
 import streamlit as st
 from apis import stock_info, gemini_technical_analysis
 from utils import session, invalid_data
+from components import support_author
 
 
 def general_ai_analysis_page():
@@ -8,7 +9,7 @@ def general_ai_analysis_page():
     symbol = session.set_session_state()
     stock = stock_info.StockInfo(symbol)
     stock_general_info = stock.get_general_info()
-    support_modal()
+    support_author.support_modal()
     if "longName" in stock_general_info:
         response_placeholder = st.empty()
         with st.spinner("Generating analysis..."):
@@ -25,32 +26,6 @@ def general_ai_analysis_page():
         response_placeholder.write(story)
     else:
         invalid_data.invalid_stock_symbol()
-
-
-@st.dialog("Support Us")
-def support_modal():
-    st.write(
-        """
-        <div style="display: flex; align-items: center;">
-            <span style="margin-right: 1rem;">❤️</span>
-            <span>
-                Support the developer: <a href="https://buymeacoffee.com/pooria7" target="_blank">Buy me a Book</a>
-            </span>
-        </div>
-        """,
-        unsafe_allow_html=True,
-    )
-    st.write(
-        """
-        <div style="display: flex; align-items: center;">
-            <span style="margin-right: 1rem;">❤️</span>
-            <span>
-                Follow the builder on X: <a href="https://x.com/PooriaTaghdiri" target="_blank">Pooria's X</a>
-            </span>
-        </div>
-        """,
-        unsafe_allow_html=True,
-    )
 
 
 general_ai_analysis_page()
